requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.9.2
pydantic-settings==2.5.2

# Database
sqlalchemy==2.0.42

# HTTP Client
httpx>=0.27.0

# AI/ML
openai>=1.0.0

# Web Search
tavily-python>=0.7.0

# Utilities
python-dotenv==1.0.0
pytz>=2024.2